+++ Assets/Scenes/Room.unity
@@ -497,6 +497,7 @@ MonoBehaviour:
     - {fileID: 687423935}
     - {fileID: 952233634}
   prefab: {fileID: 3750217307755293535, guid: 4ace620c7391ffe49b1210a909203e93, type: 3}
+  audioSource: {fileID: 607860227}
 --- !u!114 &374719061
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -870,6 +871,11 @@ PrefabInstance:
     m_AddedGameObjects: []
     m_AddedComponents: []
   m_SourcePrefab: {fileID: 100100000, guid: 2686b333b3e038448a80d3ef1e9594dc, type: 3}
+--- !u!82 &607860227 stripped
+AudioSource:
+  m_CorrespondingSourceObject: {fileID: 711034502722958836, guid: 2686b333b3e038448a80d3ef1e9594dc, type: 3}
+  m_PrefabInstance: {fileID: 607860226}
+  m_PrefabAsset: {fileID: 0}
 --- !u!1 &620808616
 GameObject:
   m_ObjectHideFlags: 0
@@ -1774,6 +1780,7 @@ GameObject:
   - component: {fileID: 1816556923}
   - component: {fileID: 1816556926}
   - component: {fileID: 1816556928}
+  - component: {fileID: 1816556929}
   m_Layer: 0
   m_Name: Canvas
   m_TagString: Untagged
@@ -1904,6 +1911,19 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: e975f0ac386b1f14380bdd996a2520b9, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+--- !u!114 &1816556929
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1816556921}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 58b22d2cc3194784da44e1f3032e88d4, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  audioSource: {fileID: 607860227}
 --- !u!1 &1916105679
 GameObject:
   m_ObjectHideFlags: 0
