+++ Assets/Scripts/ButtonSoundSetter.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 58b22d2cc3194784da44e1f3032e88d4
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
